+++ content/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "content"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+nix-docs-gen = { path = "../gen" }
+
+[[bin]]
+name = "nix-docs-content"
+path = "src/main.rs"
